First release
First working version of MxBackup

+++ mxbackup.py
@@ -19,11 +19,15 @@ import sys
 import argparse
 import csv
 import datetime
-from StringIO import StringIO
+try:
+    from StringIO import StringIO
+except ImportError:
+    from io import StringIO
 
-RELEASE = '1.0 - 3 dec 2016'
-tmpconfig = 'config.tmp'
-tmpbackup = 'backup.tmp'
+RELEASE = '1.0 - 29-8-17'
+TMPCONFIG = 'config.tmp'
+TIMEOUT = 15  # retrieving config is generally fast
+VERBOSE = 0   # show pycurl verbose
 
 class FileReader:
     def __init__(self, fp):
@@ -37,8 +41,9 @@ def filewritable(filename):
         f = open(filename, 'w')
         f.close()
     except IOError:
-        print('Unable to write to %s. It might be opened in another application. Skip operation.', filename)
+        print('Unable to write to ' + filename + '. It might be open in another application.')
         return False
+    os.remove(filename)
     return True
         
 
@@ -62,8 +67,7 @@ def transfer(ipaddr, username, password, commandfile):
     c.setopt(c.URL, 'http://' + ipaddr + '/admin/remoteconfig')
     c.setopt(c.POST, 1)
     c.setopt(c.CONNECTTIMEOUT, 5)
-    # programming a full config, update and store takes quite some time
-    c.setopt(c.TIMEOUT, 120)
+    c.setopt(c.TIMEOUT, TIMEOUT)
     filesize = os.path.getsize(commandfile)
     f = open(commandfile, 'rb')
     c.setopt(c.FAILONERROR, True)
@@ -71,7 +75,7 @@ def transfer(ipaddr, username, password, commandfile):
     c.setopt(pycurl.READFUNCTION, FileReader(f).read_callback)
     c.setopt(c.WRITEFUNCTION, storage.write)
     c.setopt(pycurl.HTTPHEADER, ["application/x-www-form-urlencoded"])
-    c.setopt(c.VERBOSE, 0)
+    c.setopt(c.VERBOSE, VERBOSE)
     c.setopt(pycurl.HTTPAUTH, pycurl.HTTPAUTH_BASIC)
     c.setopt(pycurl.USERPWD, username + ':' + password)
     try:
@@ -90,15 +94,15 @@ def transfer(ipaddr, username, password, commandfile):
 # *** Main program ***
 # ***************************************************************
 print('MxBackup ' + RELEASE + ' by (c) Simac Healthcare.')
+print('Saves entire configuration of multiple Mobotix camera\'s to local disk.')
 print('Disclaimer: ')
 print('USE THIS SOFTWARE AT YOUR OWN RISK')
 print(' ')
 
 # *** Read arguments passed on commandline
 parser = argparse.ArgumentParser()
-parser.add_argument("-v", "--verify", help="don't program camera yet but show resulting commandfile(s)", action="store_true")
-parser.add_argument("-d", "--deviceIP", nargs=1, help="specify target device IP when programming a single camera")
-parser.add_argument("-l", "--devicelist", nargs=1, help="specify target device list in CSV when programming multiple camera's")
+parser.add_argument("-d", "--deviceIP", nargs=1, help="specify target device IP when reading a single camera")
+parser.add_argument("-l", "--devicelist", nargs=1, help="specify target device list in CSV when reading multiple camera's")
 parser.add_argument("-u", "--username", nargs=1, help="specify target device admin username")
 parser.add_argument("-p", "--password", nargs=1, help="specify target device admin password")
 
@@ -120,7 +124,11 @@ if args.password is None:
     password = 'meinsm'
 else:
     password = args.password[0]
-    
+
+if not args.deviceIP and not args.devicelist:
+    print("No devices specified. Either specify a device (-d) or devicelist (-l)")
+    sys.exit()
+        
 if args.deviceIP:
     if not validate_ip(args.deviceIP[0]):
         print("The device %s is not a valid IPv4 address!" % (args.deviceIP[0]))
@@ -152,24 +160,26 @@ for devicenr in range(1, len(devicelist)):
     #skip device if starts with comment
     if devicelist[devicenr][0][0] != '#':
         # build API commandfile to read the config
-        outfile = open(tmpconfig, 'w')
-        outfile.write('helo\n')
-        outfile.write('view configfile\n')
-        outfile.write('quit\n')
-        outfile.close()
-        ipaddr = devicelist[devicenr][0]
-        cfgfilename = ipaddr.replace(".", "-") + "_" + \
-                      datetime.datetime.now().strftime("_%y%m%d-%H%M") + ".cfg"
-        if filewritable(cfgfilename):
-            (result, received) = transfer(ipaddr, username, password, tmpconfig)
-            if result:
-                print('Reading of ' + ipaddr + ' succeeded.')
-                outfile = open(cfgfilename, 'w')
-                outfile.write(received)
-                outfile.seek(-24,2)
-                outfile.truncate()
-                outfile.close()
-            else:
-                print('ERROR: Reading of ' + ipaddr + ' failed.')
+        if filewritable(TMPCONFIG):
+            outfile = open(TMPCONFIG, 'w')
+            outfile.write('helo\n')
+            outfile.write('view configfile\n')
+            outfile.write('quit\n')
+            outfile.close()
+            ipaddr = devicelist[devicenr][0]
+            cfgfilename = ipaddr.replace(".", "-") + "_" + \
+                          datetime.datetime.now().strftime("_%y%m%d-%H%M") + ".cfg"
+            if filewritable(cfgfilename):
+                (result, received) = transfer(ipaddr, username, password, TMPCONFIG)
+                if result:
+                    print('Reading of ' + ipaddr + ' succeeded.')
+                    outfile = open(cfgfilename, 'w')
+                    outfile.write(received)
+                    outfile.seek(-24,2)
+                    outfile.truncate()
+                    outfile.close()
+                else:
+                    print('ERROR: Reading of ' + ipaddr + ' failed.')
+            os.remove(TMPCONFIG)
         print('')
 print("Done.")
